+++ packages/perseus/src/widgets/radio/__tests__/radio.testdata.ts
@@ -53,24 +53,28 @@ export const questionWithPassage: PerseusRenderer = {
             options: {
                 choices: [
                     {
+                        id: "0-0-0-0-0",
                         content: "$-8$ and $8$",
                         correct: false,
                         rationale:
                             "The square root operation ($\\sqrt{\\phantom{x}}$) calculates *only* the positive square root when performed on a number, so $x$ is equal to *only* $8$.",
                     },
                     {
+                        id: "1-1-1-1-1",
                         content: "$-8$",
                         correct: false,
                         rationale:
                             "While $(-8)^2=64$, the square root operation ($\\sqrt{\\phantom{x}}$) calculates *only* the positive square root when performed on a number.",
                     },
                     {
+                        id: "2-2-2-2-2",
                         content: "$8$ {{passage-ref 1 1}}\n\n",
                         correct: true,
                         isNoneOfTheAbove: false,
                         rationale: "$8$ is the positive square root of $64$.",
                     },
                     {
+                        id: "3-3-3-3-3",
                         content: "No value of $x$ satisfies the equation.",
                         correct: false,
                         isNoneOfTheAbove: false,
@@ -120,6 +124,7 @@ export const choicesWithImages: PerseusRenderer = {
             options: {
                 choices: [
                     {
+                        id: "4-4-4-4-4",
                         content:
                             "Same \nLine\n![2 micron diameter cell](https://ka-perseus-images.s3.amazonaws.com/b17cfb6a3270c6f41f66099462e495c841cf6ca9.png)\nSame\nLine",
                         correct: false,
@@ -127,6 +132,7 @@ export const choicesWithImages: PerseusRenderer = {
                             "The markdown only has single lines between each item, so they should be treated as one complete line.",
                     },
                     {
+                        id: "5-5-5-5-5",
                         content:
                             "Text \n\nBefore\n\n![2 micron diameter cell](https://ka-perseus-images.s3.amazonaws.com/b17cfb6a3270c6f41f66099462e495c841cf6ca9.png)",
                         correct: false,
@@ -134,6 +140,7 @@ export const choicesWithImages: PerseusRenderer = {
                             "There are two 'new line' characters between the preceding text and the image. Therefore, the image should be on its own line.",
                     },
                     {
+                        id: "6-6-6-6-6",
                         content:
                             "![2 micron diameter cell](https://ka-perseus-images.s3.amazonaws.com/b17cfb6a3270c6f41f66099462e495c841cf6ca9.png)\n\nText \n\nAfter",
                         correct: false,
@@ -141,6 +148,7 @@ export const choicesWithImages: PerseusRenderer = {
                             "There are two 'new line' characters between the image and the text that follows. Therefore, the image should be on its own line.",
                     },
                     {
+                        id: "7-7-7-7-7",
                         content:
                             "Text \n\nBefore\n\n![2 micron diameter cell](https://ka-perseus-images.s3.amazonaws.com/b17cfb6a3270c6f41f66099462e495c841cf6ca9.png)\n\nText \n\nAfter",
                         correct: false,
@@ -148,6 +156,7 @@ export const choicesWithImages: PerseusRenderer = {
                             "There are two 'new line' characters between the image and the text that surrounds it. Therefore, the image should be on its own line.",
                     },
                     {
+                        id: "8-8-8-8-8",
                         content:
                             "![2 micron diameter cell](https://ka-perseus-images.s3.amazonaws.com/b17cfb6a3270c6f41f66099462e495c841cf6ca9.png)",
                         correct: false,
@@ -155,6 +164,7 @@ export const choicesWithImages: PerseusRenderer = {
                             "The markdown only has an image (no text), so nothing should be adjusted.",
                     },
                     {
+                        id: "9-9-9-9-9",
                         content:
                             "![2 micron diameter cell](https://ka-perseus-images.s3.amazonaws.com/b17cfb6a3270c6f41f66099462e495c841cf6ca9.png)\n\n![2 micron diameter cell](https://ka-perseus-images.s3.amazonaws.com/b17cfb6a3270c6f41f66099462e495c841cf6ca9.png)",
                         correct: false,
@@ -221,23 +231,27 @@ export const SingleSelectOverflowContent: PerseusRenderer = {
             options: {
                 choices: [
                     {
+                        id: "10-10-10-10-10",
                         content:
                             "$1+1+1+1+1+5+5+1+1+1+1+1+7+2+1+1+9+5+3+1+1+6+4+10+3+2$",
                         correct: true,
                         rationale: "Add the following numbers to get 75.",
                     },
                     {
+                        id: "11-11-11-11-11",
                         content: "$100-50$",
                         correct: false,
                         rationale: "Subtract the following numbers.",
                     },
                     {
+                        id: "12-12-12-12-12",
                         content: "$200-125+10$",
                         correct: false,
                         isNoneOfTheAbove: false,
                         rationale: "Calculate the following numbers.",
                     },
                     {
+                        id: "13-13-13-13-13",
                         content: "$10+10+10+10$",
                         correct: false,
                         isNoneOfTheAbove: false,