+++ packages/perseus/src/widgets/radio/multiple-choice-component.new.tsx
@@ -1,4 +1,4 @@
-import * as React from "react";
+import React, {useId} from "react";
 
 import {usePerseusI18n} from "../../components/i18n-context";
 import ScrollableView from "../../components/scrollable-view";
@@ -8,6 +8,7 @@ import styles from "./multiple-choice.module.css";
 import {getChoiceLetter} from "./util";
 import {getInstructionsText} from "./utils/string-utils";
 
+import type {IndicatorContent} from "./choice-indicator.new";
 import type {ChoiceType} from "./multiple-choice-widget.new";
 import type {PerseusStrings} from "@khanacademy/perseus/strings";
 
@@ -34,6 +35,18 @@ export interface MultipleChoiceComponentProps {
  *
  * Supports both radio button (single select) and checkbox (multiple select) modes.
  *
+ * NOTE: This widget does not implement standard radio button or checkbox elements.
+ *       The key difference is that the single-select (aka radio button) should
+ *           allow the user to deselect a choice, whereas regular radio buttons
+ *           can only be deselected by selecting another choice.
+ *       Another difference is that keyboard navigation with single-select is
+ *           done with the tab key, whereas regular radio buttons use the arrow keys,
+ *           and the radio buttons are automatically selected when navigating to them.
+ *       Since interactions with the single-select version of this widget don't
+ *           follow the standard radio button behavior, we decided to use toggle
+ *           button semantics for both versions of the widget,
+ *           and implement additional ARIA attributes where needed.
+ *
  * Created as part of the Radio Revitalization Project (LEMS-2933).
  */
 const MultipleChoiceComponent = ({
@@ -45,6 +58,7 @@ const MultipleChoiceComponent = ({
     reviewMode,
 }: MultipleChoiceComponentProps): React.ReactElement => {
     const {strings} = usePerseusI18n();
+    const legendId = useId();
 
     const instructions = getInstructionsText({
         multipleSelect,
@@ -62,9 +76,15 @@ const MultipleChoiceComponent = ({
             className={styles.container}
             data-feature-flag="feature flag is ON"
         >
-            <legend className={styles.instructions}>{instructions}</legend>
+            <legend
+                id={legendId}
+                aria-hidden="true"
+                className={styles.instructions}
+            >
+                {instructions}
+            </legend>
             <ScrollableView overflowX="auto">
-                <ul className={choiceListClasses}>
+                <ul aria-labelledby={legendId} className={choiceListClasses}>
                     <ChoiceListItems
                         choices={choices}
                         i18nStrings={strings}
@@ -92,40 +112,57 @@ interface ChoiceListItemsProps {
 const ChoiceListItems = (props: ChoiceListItemsProps): React.ReactElement => {
     const {choices, i18nStrings, multipleSelect, onChoiceChange, reviewMode} =
         props;
+    const listId = useId();
 
     const items = choices.map((choice, i) => {
         const updateChecked = (isChecked: boolean) => {
             onChoiceChange(i, isChecked);
         };
+        const contentId = `${listId}-choice-${i + 1}`;
         const choiceLetter = getChoiceLetter(i, i18nStrings);
-        const showCorrectness =
-            reviewMode === true
-                ? choice.correct
-                    ? "correct"
-                    : "wrong"
-                : undefined;
+        const srContent =
+            reviewMode && choice.correct
+                ? i18nStrings.choiceCorrect
+                : i18nStrings.choice;
+        const indicatorContent: IndicatorContent = {
+            visible: choiceLetter,
+            screenReader: srContent({letter: choiceLetter}),
+            labelledBy: contentId,
+        };
+        const showCorrectness = reviewMode
+            ? choice.correct
+                ? "correct"
+                : "wrong"
+            : undefined;
         const content = choice.isNoneOfTheAbove
             ? i18nStrings.noneOfTheAbove
             : choice.content;
-        const rationaleClasses = [styles.rationale]
-            .concat(showCorrectness === "correct" ? [styles.isCorrect] : [])
-            .join(" ");
-        const rationale =
-            reviewMode && choice.hasRationale ? (
-                <div className={rationaleClasses}>{choice.rationale}</div>
-            ) : undefined;
+        let rationale: React.ReactElement | undefined;
+        if (reviewMode && choice.hasRationale) {
+            const rationaleId = `${contentId}-rationale`;
+            indicatorContent.describedBy = rationaleId;
+            const rationaleClasses =
+                showCorrectness === "correct"
+                    ? `${styles.rationale} ${styles.isCorrect}`
+                    : styles.rationale;
+            rationale = (
+                <div id={rationaleId} className={rationaleClasses}>
+                    {choice.rationale}
+                </div>
+            );
+        }
 
         return (
             <Choice
                 key={choice.id}
                 checked={choice.checked}
-                indicatorContent={choiceLetter}
+                indicatorContent={indicatorContent}
                 isMultiSelect={multipleSelect}
                 showCorrectness={showCorrectness}
                 updateChecked={updateChecked}
             >
                 <div className={styles.content}>
-                    {content}
+                    <div id={contentId}>{content}</div>
                     {rationale}
                 </div>
             </Choice>